Minor Routing Update
Replaced challenge.js with ChallengeLayout.js

Updated routing in App.js and ExerciseLayout.js
