+++ fourbetter-app/src/components/ExerciseLayout.js
@@ -160,7 +160,7 @@ const ChatBot = ({ onClose }) => {
     return (
       <div style={containerStyle}>
         {<TextDisplay />}
-        <a href= {/exercises/+TextID+/challenges/+1}>
+        <a href= {/exercises/+TextID+/challenges/}>
         <button style={nextButtonStyle}>Challenge</button>
         </a>
         <HelpButton />
